backend/requirements.txt
# FastAPI core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Vector database
pymilvus==2.3.4

# Cache
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2

# AI/ML
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.1.1
numpy==1.24.4

# Utilities
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
